[UPDATE]修正

+++ db/migrate/20231108080848_create_posttags.rb
@@ -2,7 +2,7 @@ class CreatePosttags < ActiveRecord::Migration[6.1]
   def change
     create_table :posttags do |t|
       t.bigint :post_id, null: false
-      t.integer :tag_id, null: false
+      t.bigint :tag_id, null: false
 
       t.timestamps
     end
