+++ db/migrate/20231114083935_create_favorites.rb
@@ -1,9 +1,9 @@
 class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
-      t.integer :user_id
-      t.integer :post_image_id
-      t.integer :post_id
+      t.bigint :user_id
+      t.bigint :post_image_id
+      t.bigint :post_id
 
       t.timestamps
     end
